+++ example/lib/main.dart
@@ -18,7 +18,6 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final _cameraLinuxPlugin = CameraLinux();
   bool _isCameraOpen = false;
-  late Uint8List _latestFrameData;
   late String _base64Image;
 
   @override
@@ -32,24 +31,9 @@ class _MyAppState extends State<MyApp> {
     await _cameraLinuxPlugin.initializeCamera();
   }
 
-  // Convert Frame Into Base64
-  String uint8ListToBase64Url(Uint8List data) {
-    String base64String = base64Encode(data);
-
-    String base64Url = base64String
-        .replaceAll('+', '-')
-        .replaceAll('/', '_')
-        .replaceAll('=', '');
-
-    int requiredLength = (4 - (base64Url.length % 4)) % 4;
-    return base64Url + '=' * requiredLength;
-  }
-
-  // Capture The Frame
-  void _captureImage() {
-    _latestFrameData = _cameraLinuxPlugin.captureImage();
-    String base64Url = uint8ListToBase64Url(_latestFrameData);
-    _base64Image = base64Url;
+  // Capture The Image
+  void _captureImage() async {
+    _base64Image = await _cameraLinuxPlugin.captureImage();
     setState(() {
       _isCameraOpen = true;
     });