+++ CHANGELOG.md
@@ -1,3 +1,19 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release.
+
+## 0.0.2
+
+* Fixes in core API.
+
+## 0.0.3
+
+* Created API for storing frame into Base64 Image.
+
+## 0.0.4
+
+* Code refractor
+
+## 0.0.5
+
+* Converted API functions into asyc functions.